Xaml Generation: Move all type lookups to separate file
